requirements.txt
hdbcli
requests
python-dotenv
